+++ src/main/java/com/example/umc8th/Umc8thApplication.java
@@ -2,7 +2,9 @@ package com.example.umc8th;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
+@EnableJpaAuditing
 @SpringBootApplication
 public class Umc8thApplication {
