+++ src/main/java/com/example/umc8th/code/article/dto/ArticleRequestDTO.java
@@ -1,0 +1,12 @@
+package com.example.umc8th.code.article.dto;
+
+import lombok.Getter;
+
+public class ArticleRequestDTO {
+
+    @Getter
+    public static class CreateArticleDTO {
+        private String title;
+        private String content;
+    }
+}
