+++ src/main/java/com/example/umc8th/code/exception/BaseErrorCode.java
@@ -1,0 +1,11 @@
+package com.example.umc8th.code.exception;
+
+
+import org.springframework.http.HttpStatus;
+
+public interface BaseErrorCode {
+
+    HttpStatus getStatus();
+    String getCode();
+    String getMessage();
+}
