+++ src/main/java/com/example/umc8th/code/exception/CustomResponse.java
@@ -1,0 +1,32 @@
+package com.example.umc8th.code.exception;
+
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.annotation.JsonPropertyOrder;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import org.springframework.http.HttpStatus;
+
+
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+@JsonPropertyOrder({"isSuccess", "code", "message", "result"})
+public class CustomResponse<T> {
+
+    @JsonProperty("isSuccess")
+    private boolean isSuccess;
+
+    @JsonProperty("code")
+    private String code;
+
+    @JsonProperty("message")
+    private String message;
+
+    @JsonProperty("result")
+    private T result;
+
+
+    public static <T> CustomResponse<T> onSuccess(T result) {
+        return new CustomResponse(true, HttpStatus.OK.name(), HttpStatus.OK.getReasonPhrase(), result);
+    }
+}
+
