+++ src/main/java/com/example/umc8th/code/exception/GeneralException.java
@@ -1,0 +1,15 @@
+package com.example.umc8th.code.exception;
+
+import lombok.Getter;
+
+@Getter
+public class GeneralException extends RuntimeException{
+
+    // 예외에서 발생한 에러의 상세 내용
+    private final BaseErrorCode code;
+
+    // 생성자
+    public GeneralException(BaseErrorCode code) {
+        this.code = code;
+    }
+}
