+++ src/main/java/com/example/umc8th/config/SwaggerConfig.java
@@ -1,0 +1,35 @@
+package com.example.umc8th.config;
+
+import io.swagger.v3.oas.models.Components;
+import io.swagger.v3.oas.models.OpenAPI;
+import io.swagger.v3.oas.models.info.Info;
+import io.swagger.v3.oas.models.security.SecurityRequirement;
+import io.swagger.v3.oas.models.security.SecurityScheme;
+import io.swagger.v3.oas.models.servers.Server;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class SwaggerConfig {
+
+    @Bean
+    public OpenAPI swagger() {
+        Info info = new Info().title("Umc 8th").description("Umc 8기 실습용 Swagger").version("0.0.1");
+
+        String securityScheme = "JWT TOKEN";
+        SecurityRequirement securityRequirement = new SecurityRequirement().addList(securityScheme);
+
+        Components components = new Components()
+                .addSecuritySchemes(securityScheme, new SecurityScheme()
+                        .name(securityScheme)
+                        .type(SecurityScheme.Type.HTTP)
+                        .scheme("Bearer")
+                        .bearerFormat("JWT"));
+
+        return new OpenAPI()
+                .info(info)
+                .addServersItem(new Server().url("/"))
+                .addSecurityItem(securityRequirement)
+                .components(components);
+    }
+}
