--- src/main/java/umc/week3/code/article/repository/ArticleRepository.java
@@ -1,9 +1,0 @@
-package umc.week3.code.article.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import umc.week3.code.article.entity.Article;
-
-public interface ArticleRepository extends JpaRepository<Article, Long> {
-// JpaRepository의 첫 번째는 해당 Repository가 사용할 클래스(엔티티)가 들어가야합니다.
-// 두 번째는 id의 자료형을 적어줍니다.
-}
