--- src/main/java/umc/week3/code/article/service/command/ArticleCommandService.java
@@ -1,8 +1,0 @@
-package umc.week3.code.article.service.command;
-
-import umc.week3.code.article.dto.ArticleRequestDTO;
-import umc.week3.code.article.entity.Article;
-
-public interface ArticleCommandService {
-    Article createArticle(ArticleRequestDTO.CreateArticleDTO dto);
-}
