--- src/main/java/umc/week3/code/exception/BaseErrorCode.java
@@ -1,11 +1,0 @@
-package umc.week3.code.exception;
-
-
-import org.springframework.http.HttpStatus;
-
-public interface BaseErrorCode {
-
-    HttpStatus getStatus();
-    String getCode();
-    String getMessage();
-}
